+++ src/pages/rating_page.py
@@ -5,9 +5,10 @@ from dash.exceptions import PreventUpdate
 from dash_bootstrap_templates import ThemeSwitchAIO
 
 from main import app
+from src import analyse
 from src.configuration import config, store
 from src.graphs import rating_graph
-from src.static.static_values_enum import Format, MatchType
+from src.static.static_values_enum import Format
 from src.utils import store_util, chart_util
 
 # Define the page layout
@@ -62,26 +63,7 @@ def filter_df(account):
     if not rating_df.empty:
         rating_df.loc[:].sort_values(by='created_date', inplace=True)
 
-    result_df = pd.DataFrame()
-    if not daily_df.empty:
-        # Select Ranked battle only
-        daily_df = daily_df.loc[(daily_df.match_type == MatchType.RANKED.value)]
-
-        # Select Ranked battles and make dates on day
-        daily_df['created_date'] = pd.to_datetime(daily_df.loc[:, 'created_date']).dt.date
-
-        # First group on battle_id
-        daily_df = daily_df.groupby(['battle_id'], as_index=False).agg({'result': 'first',
-                                                                        'created_date': 'first',
-                                                                        'format': 'first'})
-        # second group on day
-        win_df = daily_df.loc[daily_df.result == 'win'].groupby(
-            ['created_date', 'result', 'format'], as_index=False).agg({'result': 'count'})
-        loss_df = daily_df.loc[daily_df.result == 'loss'].groupby(
-            ['created_date', 'result', 'format'], as_index=False).agg({'result': 'count'})
-        result_df = pd.merge(left=win_df, right=loss_df, on=['created_date', 'format'])
-        result_df.rename(columns={"result_x": "win", "result_y": "loss"}, inplace=True)
-        result_df['battles'] = result_df.win + result_df.loss
+    result_df = analyse.get_daily_battle_stats(daily_df)
 
     return rating_df.to_json(date_format='iso', orient='split'), result_df.to_json(date_format='iso', orient='split')