+++ CHANGES.md
@@ -1,7 +1,7 @@
 # Release notes
 All notable changes to this project will be documented in this file.
 
-## 0.4.0 WIP
+## 0.4.0
 - Remove previous migrations
 - Migrate battle data, add modern or wild format and re-add Zyriel secondary color.
 - Use Dash clipboard functionality iso pyperclip, remove pyperclip dependency